# Improve IPv6 parser performance/simplicity