--- a/test/Test.hs
+++ b/test/Test.hs
@@ -225,19 +225,36 @@ testIPv6ParserFailure = do
   go "1:::"
   go ":1::"
   go "::1:"
-  go "1:2:3:4:5:6:7:8:"
-  go ":1:2:3:4:5:6:7:8"
+  go "1:2:3:4:5:6:777:8:"
+  go ":1:2:3:4:5:6:7777:8"
 
   -- Incorrect numbers of parts:
-  go "1111:2222:3333:4444:5555:6666::7777:8888"
+  go ""
+  go "1111"
+  go "1111:2222"
+  go "1111:2222:3333"
+  go "1111:2222:3333:4444"
+  go "1111:2222:3333:4444:5555"
+  go "1111:2222:3333:4444:5555:6666"
+  go "1111:2222:3333:4444:5555:6666:7777"
   go "1111:2222:3333:4444:5555:6666:7777:8888:9999"
-  go "1111:2222:3333:4444:5555:6666:7777:8888::9999"
+
+  -- Incorrect use of double-colon:
+  go "1111::2222::3333"
+  go "1111:2222:3333:4444:5555:6666::7777:8888" -- not needed
+  go "1111:2222:3333:4444:5555:6666:7777:8888::9999" -- too long
 
   -- IPv4 decimal embedded, with not enough parts:
   go "1:127.0.0.1"
   go "1:2:3:127.0.0.1"
   go "1:2:3:4:127.0.0.1"
   go "1:2:3:4:5:127.0.0.1"
+
+  -- IPv4 decimal before double-colon:
+  go "1:127.0.0.1::"
+
+  -- Only IPv4:
+  go "127.0.0.1"
 
   -- IPv4 decimal embedded, with too many parts:
   go "1:2:3:4:5:6:7:127.0.0.1"
@@ -248,8 +265,7 @@ testIPv6ParserFailure = do
     @=? bimap (\_ -> ()) HexIPv6
       (AT.parseOnly
         (IPv6.parser <* AT.endOfInput)
-        (Text.pack str)
-      )
+        (Text.pack str))
 
 testIPv6Encode :: Assertion
 testIPv6Encode = do
@@ -294,7 +310,7 @@ testIPv6Encode = do
    roundTripsTo s sExpected =
      case AT.parseOnly (IPv6.parser <* AT.endOfInput) (Text.pack s) of
         Right result -> IPv6.encode result @?= Text.pack sExpected
-        Left failMsg -> fail failMsg -- parse shouldn't fail here
+        Left failMsg -> fail ("failed to parse '" ++ s ++ "': " ++ failMsg)
 
 textBadIPv4 :: [String]
 textBadIPv4 =