Since performance was slightly worsened by #37, I did some work to both simplify the IPv6 parser for understandability, and to make it faster.

All the main cases are now faster than the pre-#37 parser.

Here are the benchmark differences for this PR:

```
benchmarking IPv6 from Text/New '::'
time                 94.52 ns   (92.94 ns .. 96.27 ns)
                     0.996 R²   (0.994 R² .. 0.998 R²)
mean                 97.46 ns   (95.18 ns .. 102.9 ns)
std dev              11.55 ns   (7.396 ns .. 19.11 ns)
variance introduced by outliers: 93% (severely inflated)

benchmarking IPv6 from Text/Old '::'
time                 106.2 ns   (104.4 ns .. 108.9 ns)
                     0.995 R²   (0.991 R² .. 0.998 R²)
mean                 109.2 ns   (106.6 ns .. 113.9 ns)
std dev              11.28 ns   (7.229 ns .. 16.38 ns)
variance introduced by outliers: 91% (severely inflated)

benchmarking IPv6 from Text/New '1:2:3:4:5:6:7:8'
time                 570.5 ns   (566.2 ns .. 576.0 ns)
                     0.998 R²   (0.995 R² .. 0.999 R²)
mean                 594.5 ns   (580.0 ns .. 630.7 ns)
std dev              75.65 ns   (45.32 ns .. 124.5 ns)
variance introduced by outliers: 93% (severely inflated)

benchmarking IPv6 from Text/Old '1:2:3:4:5:6:7:8'
time                 954.7 ns   (946.1 ns .. 963.0 ns)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 949.7 ns   (944.2 ns .. 955.7 ns)
std dev              19.29 ns   (16.17 ns .. 23.61 ns)
variance introduced by outliers: 24% (moderately inflated)

benchmarking IPv6 from Text/New '1:2::7:8'
time                 438.6 ns   (435.6 ns .. 441.9 ns)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 439.3 ns   (436.9 ns .. 441.6 ns)
std dev              7.777 ns   (6.695 ns .. 9.178 ns)
variance introduced by outliers: 21% (moderately inflated)

benchmarking IPv6 from Text/Old '1:2::7:8'
time                 547.8 ns   (543.4 ns .. 552.9 ns)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 547.1 ns   (543.8 ns .. 553.6 ns)
std dev              14.48 ns   (9.083 ns .. 27.68 ns)
variance introduced by outliers: 36% (moderately inflated)

benchmarking IPv6 from Text/New 'a:b::c:d'
time                 371.7 ns   (368.4 ns .. 375.7 ns)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 370.5 ns   (368.2 ns .. 373.3 ns)
std dev              8.435 ns   (6.527 ns .. 12.45 ns)
variance introduced by outliers: 30% (moderately inflated)

benchmarking IPv6 from Text/Old 'a:b::c:d'
time                 428.0 ns   (425.4 ns .. 431.2 ns)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 428.8 ns   (426.9 ns .. 430.7 ns)
std dev              6.405 ns   (5.121 ns .. 8.273 ns)
variance introduced by outliers: 15% (moderately inflated)
```