@andrewthad 2018-06-26
I'll take a closer look at this in a little bit. One thing that sticks out to me is that there are top-level functions that don't have type signatures. Type signatures would improve the readability of the implementation.