@andrewthad 2018-07-16
I realized that those weren't actually top-level functions I was looking at, but I've added type signatures to them anyway. I cannot see any obvious way to convince GHC the the pattern match on the list is always going to match the 8-element case (indeed, I cannot really even convince myself of this), but I've added a more information error message in case this happens so that people will know to report it. Merging.